pyproject.toml
[project]
name = "ai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastmcp[full]>=2.13.1",
    "flask>=3.1.2",
    "google-generativeai",
    "langchain",
    "langchain-google-genai>=2.0.10",
    "langchain-mcp-adapters>=0.1.14",
    "langgraph",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-dotenv",
]
